The following changes were made to the project: a. Added a new file called jwtGenerator.js which generates a JWT token for the user. b. Added a new file called authentication_middleware.js which is a middleware that checks if the user is authenticated or not. c. Added a new route called /login which is used to login the user and generate a JWT token for the user. d. Added a new route called /users which is used to get all the users in the database. e. Modified the gitignore file to ignore .vscode folder. f. Modified the README.md file to add the changes made to the project. g. Modified the server.js file to add the new routes and middlewares.
The use of stateless authentication has been implemented in the project. The JWT token is generated when the user logs in and is sent to the client. The client then sends the JWT token in the header of the request to the server. The server then verifies the JWT token and if the token is valid then the user is authenticated and the request is processed. If the token is invalid then the user is not authenticated and the request is not processed.